diary_chatbot/requirements.txt
openai
streamlit
numpy
pandas
python-dotenv
faiss-cpu
sentence-transformers